Add support for fully-predicated loops
This patch adds support for one of the main features of SVE: the ability
to use a single fully-predicated loop (here called "fully-masked" to
follow existing GCC terminology).  A WHILE instruction generates the
predicate/mask for each iteration of the loop, given the current scalar
iv value and the loop bound.

Using full masking requires new optabs and internal functions for
masked arithmetic operations.  These are particularly useful when doing
reductions, so that we don't include inactive vector elements in the
accumulated result.